+++ Events/Subscribers.cs
@@ -2,6 +2,8 @@
 using OpenLib.Common;
 using ShipColors.ConfigManager;
 using ShipColors.Customizer;
+using System.Reflection;
+using System;
 
 
 namespace ShipColors.Events
@@ -11,7 +13,7 @@ namespace ShipColors.Events
         public static void Subscribe()
         {
             EventManager.TerminalAwake.AddListener(OnTerminalAwake);
-            EventManager.StartOfRoundAwake.AddListener(OnStart);
+            EventManager.StartOfRoundStart.AddListener(OnStart);
             EventManager.TerminalDisable.AddListener(OnTerminalDisable);
             EventManager.GameNetworkManagerStart.AddListener(OnGameLoad);
         }
@@ -20,12 +22,22 @@ namespace ShipColors.Events
         {
             if(StartGame.SoftCompatibility("darmuh.TerminalStuff", ref Plugin.instance.darmuhsTerminalStuff))
             {
-                Plugin.MoreLogs("leaving terminal customization to darmuhsTerminalStuff");
+                Plugin.MoreLogs("leaving terminal customization to darmuhsTerminalStuff (generated config will skip terminal object)");
             }
             if(StartGame.SoftCompatibility("BMX.LobbyCompatibility", ref Plugin.instance.LobbyCompat))
             {
-                Compat.BMX_LobbyCompat.SetCompat(false);
+                Version version = Assembly.GetExecutingAssembly().GetName().Version;
+                OpenLib.Compat.BMX_LobbyCompat.SetBMXCompat(false, version);
             }
+
+            if (OpenLib.Plugin.instance.LethalConfig)
+            {
+                OpenLib.Compat.LethalConfigSoft.AddButton("Setup", "Generate Webpage", "Press this to generate a webpage in the Bepinex/Config/Webconfig folder from this mod's generated config!\nYou can then use this webpage to modify your config and paste a config code to apply in-game", "Generate Webpage", GeneratedConfig.GenerateWebpage);
+                OpenLib.Compat.LethalConfigSoft.AddButton("Setup", "Regen Config", "Press this to regenerate the generated config when [Mode Setting] is set to Generate Config.", "Regen Config", GeneratedConfig.RegenerateConfig);
+            }
+            else
+                Plugin.MoreLogs("LethalConfig is not detected by OpenLib");
+            
         }
 
         public static void OnTerminalAwake(Terminal instance)
@@ -41,11 +53,23 @@ namespace ShipColors.Events
                 CustomShipLights.SetShipLights();
             }
 
-            if (ConfigSettings.UseSharedMaterials.Value)
+            if (ConfigSettings.ModeSetting.Value == "Use Shared Textures")
             {
                 GlobalSharedCustomization.UseSharedTextures();
                 Plugin.Spam("Only setting shared texture values");
                 return;
+            }
+            else if(ConfigSettings.ModeSetting.Value == "Generate Config")
+            {
+                if (!GeneratedCustomization.configGenerated)
+                {
+                    GeneratedCustomization.CreateAllConfigs();
+                    GeneratedConfig.Generated.SettingChanged += GeneratedConfig.OnSettingChanged;
+                    GeneratedConfig.ReadConfigCode();
+                }
+                    
+                else
+                    GeneratedCustomization.ReadCustomClassValues(ref GeneratedCustomization.materialToColor);
             }
             
             //TerminalCustomizations.TerminalStuff();
@@ -53,7 +77,8 @@ namespace ShipColors.Events
 
         public static void OnTerminalDisable()
         {
-            //nothing needed here
+            GeneratedCustomization.materialToColor.Clear();
+            GeneratedCustomization.configGenerated = false;
         }
 
         public static void OnStart()